+++ libs/vkd3d/state.c
@@ -3227,232 +3227,3 @@ bool vkd3d_bindless_state_find_binding(const struct vkd3d_bindless_state *bindle
 
     return false;
 }
-
-/* ID3D12PipelineLibrary */
-static inline struct d3d12_pipeline_library *impl_from_ID3D12PipelineLibrary(d3d12_pipeline_library_iface *iface)
-{
-    return CONTAINING_RECORD(iface, struct d3d12_pipeline_library, ID3D12PipelineLibrary_iface);
-}
-
-static HRESULT STDMETHODCALLTYPE d3d12_pipeline_library_QueryInterface(d3d12_pipeline_library_iface *iface,
-        REFIID riid, void **object)
-{
-    TRACE("iface %p, riid %s, object %p.\n", iface, debugstr_guid(riid), object);
-
-    if (IsEqualGUID(riid, &IID_ID3D12PipelineLibrary)
-            || IsEqualGUID(riid, &IID_ID3D12PipelineLibrary1)
-            || IsEqualGUID(riid, &IID_ID3D12DeviceChild)
-            || IsEqualGUID(riid, &IID_ID3D12Object)
-            || IsEqualGUID(riid, &IID_IUnknown))
-    {
-        ID3D12PipelineLibrary_AddRef(iface);
-        *object = iface;
-        return S_OK;
-    }
-
-    WARN("%s not implemented, returning E_NOINTERFACE.\n", debugstr_guid(riid));
-
-    *object = NULL;
-    return E_NOINTERFACE;
-}
-
-static ULONG STDMETHODCALLTYPE d3d12_pipeline_library_AddRef(d3d12_pipeline_library_iface *iface)
-{
-    struct d3d12_pipeline_library *pipeline_library = impl_from_ID3D12PipelineLibrary(iface);
-    ULONG refcount = InterlockedIncrement(&pipeline_library->refcount);
-
-    TRACE("%p increasing refcount to %u.\n", pipeline_library, refcount);
-
-    return refcount;
-}
-
-static ULONG STDMETHODCALLTYPE d3d12_pipeline_library_Release(d3d12_pipeline_library_iface *iface)
-{
-    struct d3d12_pipeline_library *pipeline_library = impl_from_ID3D12PipelineLibrary(iface);
-    ULONG refcount = InterlockedDecrement(&pipeline_library->refcount);
-
-    TRACE("%p decreasing refcount to %u.\n", pipeline_library, refcount);
-
-    if (!refcount)
-    {
-        struct d3d12_device *device = pipeline_library->device;
-        vkd3d_private_store_destroy(&pipeline_library->private_store);
-        vkd3d_free(pipeline_library);
-        d3d12_device_release(device);
-    }
-
-    return refcount;
-}
-
-static HRESULT STDMETHODCALLTYPE d3d12_pipeline_library_GetPrivateData(d3d12_pipeline_library_iface *iface,
-        REFGUID guid, UINT *data_size, void *data)
-{
-    struct d3d12_pipeline_library *pipeline_library = impl_from_ID3D12PipelineLibrary(iface);
-
-    TRACE("iface %p, guid %s, data_size %p, data %p.\n", iface, debugstr_guid(guid), data_size, data);
-
-    return vkd3d_get_private_data(&pipeline_library->private_store, guid, data_size, data);
-}
-
-static HRESULT STDMETHODCALLTYPE d3d12_pipeline_library_SetPrivateData(d3d12_pipeline_library_iface *iface,
-        REFGUID guid, UINT data_size, const void *data)
-{
-    struct d3d12_pipeline_library *pipeline_library = impl_from_ID3D12PipelineLibrary(iface);
-
-    TRACE("iface %p, guid %s, data_size %u, data %p.\n", iface, debugstr_guid(guid), data_size, data);
-
-    return vkd3d_set_private_data(&pipeline_library->private_store, guid, data_size, data);
-}
-
-static HRESULT STDMETHODCALLTYPE d3d12_pipeline_library_SetPrivateDataInterface(d3d12_pipeline_library_iface *iface,
-        REFGUID guid, const IUnknown *data)
-{
-    struct d3d12_pipeline_library *pipeline_library = impl_from_ID3D12PipelineLibrary(iface);
-
-    TRACE("iface %p, guid %s, data %p.\n", iface, debugstr_guid(guid), data);
-
-    return vkd3d_set_private_data_interface(&pipeline_library->private_store, guid, data);
-}
-
-static HRESULT STDMETHODCALLTYPE d3d12_pipeline_library_SetName(d3d12_pipeline_library_iface *iface, const WCHAR *name)
-{
-    struct d3d12_pipeline_library *pipeline_library = impl_from_ID3D12PipelineLibrary(iface);
-
-    TRACE("iface %p, name %s.\n", iface, debugstr_w(name, pipeline_library->device->wchar_size));
-
-    return name ? S_OK : E_INVALIDARG;
-}
-
-static HRESULT STDMETHODCALLTYPE d3d12_pipeline_library_GetDevice(d3d12_pipeline_library_iface *iface,
-        REFIID iid, void **device)
-{
-    struct d3d12_pipeline_library *pipeline_library = impl_from_ID3D12PipelineLibrary(iface);
-
-    TRACE("iface %p, iid %s, device %p.\n", iface, debugstr_guid(iid), device);
-
-    return d3d12_device_query_interface(pipeline_library->device, iid, device);
-}
-
-static HRESULT STDMETHODCALLTYPE d3d12_pipeline_library_StorePipeline(d3d12_pipeline_library_iface *iface,
-        LPCWSTR name, ID3D12PipelineState *pipeline)
-{
-    struct d3d12_pipeline_library *pipeline_library = impl_from_ID3D12PipelineLibrary(iface);
-
-    FIXME("iface %p, name %s, pipeline %p stub!\n", iface, debugstr_w(name, pipeline_library->device->wchar_size), pipeline);
-
-    return S_OK;
-}
-
-static HRESULT STDMETHODCALLTYPE d3d12_pipeline_library_LoadGraphicsPipeline(d3d12_pipeline_library_iface *iface,
-        LPCWSTR name, const D3D12_GRAPHICS_PIPELINE_STATE_DESC *desc, REFIID iid, void **pipeline_state)
-{
-    struct d3d12_pipeline_library *pipeline_library = impl_from_ID3D12PipelineLibrary(iface);
-
-    FIXME("iface %p, name %s, desc %p, iid %s, pipeline_state %p stub!\n", iface,
-            debugstr_w(name, pipeline_library->device->wchar_size),
-            desc, debugstr_guid(iid), pipeline_state);
-
-    return E_INVALIDARG;
-}
-
-static HRESULT STDMETHODCALLTYPE d3d12_pipeline_library_LoadComputePipeline(d3d12_pipeline_library_iface *iface,
-        LPCWSTR name, const D3D12_COMPUTE_PIPELINE_STATE_DESC *desc, REFIID iid, void **pipeline_state)
-{
-    struct d3d12_pipeline_library *pipeline_library = impl_from_ID3D12PipelineLibrary(iface);
-
-    FIXME("iface %p, name %s, desc %p, iid %s, pipeline_state %p stub!\n", iface,
-            debugstr_w(name, pipeline_library->device->wchar_size),
-            desc, debugstr_guid(iid), pipeline_state);
-
-    return E_INVALIDARG;
-}
-
-static SIZE_T STDMETHODCALLTYPE d3d12_pipeline_library_GetSerializedSize(d3d12_pipeline_library_iface *iface)
-{
-    FIXME("iface %p stub!\n", iface);
-
-    return 0;
-}
-
-static HRESULT STDMETHODCALLTYPE d3d12_pipeline_library_Serialize(d3d12_pipeline_library_iface *iface,
-        void *data, SIZE_T data_size)
-{
-    FIXME("iface %p, data %p, data_size %lu stub!\n", iface, data, data_size);
-
-    return E_NOTIMPL;
-}
-
-static HRESULT STDMETHODCALLTYPE d3d12_pipeline_library_LoadPipeline(d3d12_pipeline_library_iface *iface,
-        LPCWSTR name, const D3D12_PIPELINE_STATE_STREAM_DESC *desc, REFIID iid, void **pipeline_state)
-{
-    struct d3d12_pipeline_library *pipeline_library = impl_from_ID3D12PipelineLibrary(iface);
-
-    FIXME("iface %p, name %s, desc %p, iid %s, pipeline_state %p stub!\n", iface,
-            debugstr_w(name, pipeline_library->device->wchar_size),
-            desc, debugstr_guid(iid), pipeline_state);
-
-    return E_INVALIDARG;
-}
-
-static CONST_VTBL struct ID3D12PipelineLibrary1Vtbl d3d12_pipeline_library_vtbl =
-{
-    /* IUnknown methods */
-    d3d12_pipeline_library_QueryInterface,
-    d3d12_pipeline_library_AddRef,
-    d3d12_pipeline_library_Release,
-    /* ID3D12Object methods */
-    d3d12_pipeline_library_GetPrivateData,
-    d3d12_pipeline_library_SetPrivateData,
-    d3d12_pipeline_library_SetPrivateDataInterface,
-    d3d12_pipeline_library_SetName,
-    /* ID3D12DeviceChild methods */
-    d3d12_pipeline_library_GetDevice,
-    /* ID3D12PipelineLibrary methods */
-    d3d12_pipeline_library_StorePipeline,
-    d3d12_pipeline_library_LoadGraphicsPipeline,
-    d3d12_pipeline_library_LoadComputePipeline,
-    d3d12_pipeline_library_GetSerializedSize,
-    d3d12_pipeline_library_Serialize,
-    /* ID3D12PipelineLibrary1 methods */
-    d3d12_pipeline_library_LoadPipeline,
-};
-
-static HRESULT d3d12_pipeline_library_init(struct d3d12_pipeline_library *pipeline_library,
-        struct d3d12_device *device, const void *blob, size_t blob_length)
-{
-    HRESULT hr;
-
-    memset(pipeline_library, 0, sizeof(*pipeline_library));
-    pipeline_library->ID3D12PipelineLibrary_iface.lpVtbl = &d3d12_pipeline_library_vtbl;
-    pipeline_library->refcount = 1;
-
-    if (FAILED(hr = vkd3d_private_store_init(&pipeline_library->private_store)))
-        goto fail;
-
-    d3d12_device_add_ref(pipeline_library->device = device);
-    return S_OK;
-
-fail:
-    return hr;
-}
-
-HRESULT d3d12_pipeline_library_create(struct d3d12_device *device, const void *blob,
-        size_t blob_length, struct d3d12_pipeline_library **pipeline_library)
-{
-    struct d3d12_pipeline_library *object;
-    HRESULT hr;
-
-    if (!(object = vkd3d_malloc(sizeof(*object))))
-        return E_OUTOFMEMORY;
-
-    if (FAILED(hr = d3d12_pipeline_library_init(object, device, blob, blob_length)))
-    {
-        vkd3d_free(object);
-        return hr;
-    }
-
-    TRACE("Created pipeline library %p.\n", object);
-
-    *pipeline_library = object;
-    return S_OK;
-}
